Reward drawer grouping
Added accordion functionality to the reward drawer so that each section is in its own. Updated the z-indexing for the reward and discount drawer as they're still showing behind the getting started button. Move the eligible partners to within the same accordion as the eligibility selection. This makes the content flow better. If a user has selected that all partners get the reward, then the next time that they open the drawer, that accordion is closed by default as it's not something you can change and you can't edit anyways. If the user selects a specific partner, then it will be open.